to attribute compositions
You can still custom register methods with the (new) MethodComposition type, but if you are developing a standalone DLL you can also simply use the Composition and (optional) CompositionDescription attributes.

This should make it much easier to manage adding and removing listener methods, as currently a lot of callback code is required.